+++ common/oidn_utils.cpp
@@ -53,7 +53,7 @@ OIDN_NAMESPACE_BEGIN
     default:
       throw std::invalid_argument("invalid device type");
     }
-    
+
     return sm;
   }
 
@@ -76,6 +76,19 @@ OIDN_NAMESPACE_BEGIN
     else
       throw std::invalid_argument("invalid device type");
 
+    return sm;
+  }
+
+  std::ostream& operator <<(std::ostream& sm, Quality quality)
+  {
+    switch (quality)
+    {
+    case Quality::Default:  sm << "default";  break;
+    case Quality::High:     sm << "high";     break;
+    case Quality::Balanced: sm << "balanced"; break;
+    default:
+      throw std::invalid_argument("invalid quality mode");
+    }
     return sm;
   }