+++ common/oidn_utils.h
@@ -17,6 +17,8 @@ OIDN_NAMESPACE_BEGIN
   std::ostream& operator <<(std::ostream& sm, DeviceType deviceType);
   std::istream& operator >>(std::istream& sm, DeviceType& deviceType);
 
+  std::ostream& operator <<(std::ostream& sm, Quality quality);
+
   std::ostream& operator <<(std::ostream& sm, const UUID& uuid);
   std::ostream& operator <<(std::ostream& sm, const LUID& luid);