+++ devices/cuda/cutlass_conv_sm70.cu
@@ -13,9 +13,11 @@ std::vector<CutlassConvFactory> getCutlassConvInstances<70>()
   using cutlass::gemm::GemmShape;
 
   return {
-    // Use float accumulation because precision is too low with half and Volta's smaller MMA shape
     CutlassConvInstance<half, float, Sm70, GemmShape<256, 32, 32>, GemmShape<64, 32, 32>, 2>::get(),
     CutlassConvInstance<half, float, Sm70, GemmShape<256, 64, 32>, GemmShape<64, 64, 32>, 2>::get(),
+
+    CutlassConvInstance<half, half,  Sm70, GemmShape<256, 32, 32>, GemmShape<64, 32, 32>, 2>::get(),
+    CutlassConvInstance<half, half,  Sm70, GemmShape<256, 64, 32>, GemmShape<64, 64, 32>, 2>::get(),
   };
 }