+++ devices/cuda/cutlass_conv_sm80.cu
@@ -13,8 +13,11 @@ std::vector<CutlassConvFactory> getCutlassConvInstances<80>()
   using cutlass::gemm::GemmShape;
 
   return {
-    CutlassConvInstance<half, half, Sm80, GemmShape<256, 32, 32>, GemmShape<64, 32, 32>, 3 /*4*/>::get(),
-    CutlassConvInstance<half, half, Sm80, GemmShape<256, 64, 32>, GemmShape<64, 64, 32>, 3>::get(),
+    CutlassConvInstance<half, float, Sm80, GemmShape<256, 32, 32>, GemmShape<64, 32, 32>, 3 /*4*/>::get(),
+    CutlassConvInstance<half, float, Sm80, GemmShape<256, 64, 32>, GemmShape<64, 64, 32>, 3>::get(),
+
+    CutlassConvInstance<half, half,  Sm80, GemmShape<256, 32, 32>, GemmShape<64, 32, 32>, 3 /*4*/>::get(),
+    CutlassConvInstance<half, half,  Sm80, GemmShape<256, 64, 32>, GemmShape<64, 64, 32>, 3>::get(),
   };
 }